Fix Small Artifact Problem
By using the desired height, I would sometimes get a small gap between
elements when placing them. By switching to a set size in the table and
passing that size down to the Canvas, we can measure and arrange the
elements by that size. This gets rid of that small gap.