Switched to dot Net 4.5
I don't know VS 2012 created the WPF application on the dot Net 4.0
framework, but this was supposed to be on the dot Net 4.5 framework from
the very start. I also removed a line of code from TestDemo01.csproj
that was removing the debug information from the application. I don't
know why VS felt like putting it in, but taking it make working on this
a lot nicer.