Fixed the Artifacting in Issue #4
I fixed the artifacting that was going over the border, but there's
still a bit of artifacting on the glow effect for the top of the
chemical name. The very top of the blur effect for the text cuts off
suddenly, making it look a bit like a bright spot there. It's not as
nice as I would like it, but it is a separate issue and good enough for
now.